pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["labidahrom <labidahrom@yandex.ru>"]
readme = "README.md"
packages = [
{include = "gendiff"}
]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.3"
flake8 = "^5.0.4"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"
argcomplete = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
just = "gendiff.scripts.just:main"
